Uncontracted first derivative of inte1e_grid working
