+++ gpu4pyscf/lib/gint/CMakeLists.txt
@@ -26,6 +26,7 @@ add_library(gint SHARED
   constant.cu
   nr_fill_ao_ints.cu
   nr_fill_ao_int3c1e.cu
+  nr_fill_ao_int3c1e_ip.cu
   nr_fill_ao_int3c2e.cu
   nr_fill_ao_int3c2e_ip1.cu
   nr_fill_ao_int3c2e_ip2.cu
