+++ gpu4pyscf/lib/gint/g3c1e.cu
@@ -62,9 +62,9 @@ static void GINTwrite_int3c1e(const double* g, double* output, const int ish, co
 
 template <int NROOTS, int GSIZE_INT3C_1E>
 __global__
-void GINTfill_int3c1e_kernel_general(double* output, const BasisProdOffsets offsets, const int i_l, const int j_l, const int nprim_ij,
-                                     const int stride_j, const int stride_ij, const int ao_offsets_i, const int ao_offsets_j,
-                                     const double omega, const double* grid_points, const double* charge_exponents)
+static void GINTfill_int3c1e_kernel_general(double* output, const BasisProdOffsets offsets, const int i_l, const int j_l, const int nprim_ij,
+                                            const int stride_j, const int stride_ij, const int ao_offsets_i, const int ao_offsets_j,
+                                            const double omega, const double* grid_points, const double* charge_exponents)
 {
     const int ntasks_ij = offsets.ntasks_ij;
     const int ngrids = offsets.ntasks_kl;
@@ -87,7 +87,7 @@ void GINTfill_int3c1e_kernel_general(double* output, const BasisProdOffsets offs
     double g[GSIZE_INT3C_1E];
 
     for (int ij = prim_ij; ij < prim_ij+nprim_ij; ++ij) {
-        GINTg1e<NROOTS>(g, grid_point, ish, jsh, ij, i_l, j_l, charge_exponent, omega);
+        GINT_g1e<NROOTS>(g, grid_point, ish, jsh, ij, i_l, j_l, charge_exponent, omega);
         GINTwrite_int3c1e<NROOTS>(g, output, ish, jsh, task_grid, i_l, j_l, stride_j, stride_ij, ao_offsets_i, ao_offsets_j);
     }
 }
@@ -131,9 +131,9 @@ static void GINTwrite_int3c1e_charge_contracted(const double* g, double* local_o
 
 template <int NROOTS, int GSIZE_INT3C_1E>
 __global__
-void GINTfill_int3c1e_charge_contracted_kernel_general(double* output, const BasisProdOffsets offsets, const int i_l, const int j_l, const int nprim_ij,
-                                                       const int stride_j, const int stride_ij, const int ao_offsets_i, const int ao_offsets_j,
-                                                       const double omega, const double* grid_points, const double* charge_exponents)
+static void GINTfill_int3c1e_charge_contracted_kernel_general(double* output, const BasisProdOffsets offsets, const int i_l, const int j_l, const int nprim_ij,
+                                                              const int stride_j, const int stride_ij, const int ao_offsets_i, const int ao_offsets_j,
+                                                              const double omega, const double* grid_points, const double* charge_exponents)
 {
     const int ntasks_ij = offsets.ntasks_ij;
     const int ngrids = offsets.ntasks_kl;
@@ -163,7 +163,7 @@ void GINTfill_int3c1e_charge_contracted_kernel_general(double* output, const Bas
         double g[GSIZE_INT3C_1E];
 
         for (int ij = prim_ij; ij < prim_ij+nprim_ij; ++ij) {
-            GINTg1e<NROOTS>(g, grid_point, ish, jsh, ij, i_l, j_l, charge_exponent, omega);
+            GINT_g1e<NROOTS>(g, grid_point, ish, jsh, ij, i_l, j_l, charge_exponent, omega);
             GINTwrite_int3c1e_charge_contracted<NROOTS>(g, output_cache, charge, i_l, j_l);
         }
     }
@@ -182,9 +182,9 @@ void GINTfill_int3c1e_charge_contracted_kernel_general(double* output, const Bas
 
 template <int NROOTS>
 __global__
-void GINTfill_int3c1e_density_contracted_kernel_general(double* output, const double* density, const HermiteDensityOffsets hermite_density_offsets,
-                                                        const BasisProdOffsets offsets, const int i_l, const int j_l, const int nprim_ij,
-                                                        const double omega, const double* grid_points, const double* charge_exponents)
+static void GINTfill_int3c1e_density_contracted_kernel_general(double* output, const double* density, const HermiteDensityOffsets hermite_density_offsets,
+                                                               const BasisProdOffsets offsets, const int i_l, const int j_l, const int nprim_ij,
+                                                               const double omega, const double* grid_points, const double* charge_exponents)
 {
     const int ntasks_ij = offsets.ntasks_ij;
     const int ngrids = offsets.ntasks_kl;
